Created MarkovChain class.
I realized I was manipulating the ACTUAL defaultdict of the Markov chain with handlers and stuff that I should just go ahead and subclass the chain from defaultdict and stuff all the logic I can there. There's quite a bit of mucking about with internals in the chain class. And type checking (sorry).